Disable urngd as well
The urngd of the OpenWRT, acting as an EGD, does not work well when OpenWRT running as container guest, and may occasionally cause high usage of CPU assigned to the instance when low on entropy. Just disable it, leave the task to the EGD of the host, as the OpenWRT container instances presumably share entropy with the host.